Begun rdp command builder

+++ n64-sys/src/ai.rs
@@ -7,27 +7,27 @@ use core::ptr::{read_volatile, write_volatile};
 const AI_BASE: usize = 0xA4500000;
 
 const AI_ADDR: *mut usize = (AI_BASE + 0x00) as _;
-const AI_LENGTH: *mut u32 = (AI_BASE + 0x04) as _;
-const AI_CONTROL: *mut u32 = (AI_BASE + 0x08) as _;
-const AI_STATUS: *mut u32 = (AI_BASE + 0x0C) as _;
-const AI_DACRATE: *mut u32 = (AI_BASE + 0x10) as _;
-const AI_SAMPLESIZE: *mut u32 = (AI_BASE + 0x14) as _;
+const AI_LENGTH: *mut usize = (AI_BASE + 0x04) as _;
+const AI_CONTROL: *mut usize = (AI_BASE + 0x08) as _;
+const AI_STATUS: *mut usize = (AI_BASE + 0x0C) as _;
+const AI_DACRATE: *mut usize = (AI_BASE + 0x10) as _;
+const AI_SAMPLESIZE: *mut usize = (AI_BASE + 0x14) as _;
 
-const AI_NTSC_DACRATE: u32 = 48681812;
-const AI_PAL_DACRATE: u32 = 49656530;
-const AI_MPAL_DACRATE: u32 = 48628316;
+const AI_NTSC_DACRATE: usize = 48681812;
+const AI_PAL_DACRATE: usize = 49656530;
+const AI_MPAL_DACRATE: usize = 48628316;
 
-const AI_STATUS_BUSY: u32 = 1 << 30;
-const AI_STATUS_FULL: u32 = 1 << 31;
+const AI_STATUS_BUSY: usize = 1 << 30;
+const AI_STATUS_FULL: usize = 1 << 31;
 
 const TV_TYPE_LOC: usize = 0x80000300;
 
-const FREQUENCY: u32 = 22050;
+const FREQUENCY: usize = 22050;
 const BUFFER_COUNT: usize = 4;
 
 pub const BUFFER_NO_SAMPLES: usize = 2 * 512;
 
-static mut REAL_FREQUENCY: u32 = 0;
+static mut REAL_FREQUENCY: usize = 0;
 static mut BUFFERS: [[i16; BUFFER_NO_SAMPLES]; BUFFER_COUNT] =
     [[0; BUFFER_NO_SAMPLES]; BUFFER_COUNT];
 
@@ -48,7 +48,7 @@ fn ai_full() -> bool {
 #[inline]
 pub fn init() {
     unsafe {
-        let clockrate = match read_volatile(TV_TYPE_LOC as *const u32) {
+        let clockrate = match read_volatile(TV_TYPE_LOC as *const usize) {
             0 => AI_PAL_DACRATE,
             2 => AI_MPAL_DACRATE,
             _ => AI_NTSC_DACRATE,
@@ -63,7 +63,6 @@ pub fn init() {
 
 #[inline]
 pub fn write_audio_blocking(buffer: &[i16]) {
-
     if buffer.len() != BUFFER_NO_SAMPLES {
         panic!();
     }
@@ -114,7 +113,7 @@ pub fn submit_audio_data_to_dac() {
                 virtual_to_physical(uncached_addr(BUFFERS[NOW_PLAYING].as_ptr())),
             );
             memory_barrier();
-            write_volatile(AI_LENGTH, ((BUFFER_NO_SAMPLES * 2) & !7) as u32);
+            write_volatile(AI_LENGTH, (BUFFER_NO_SAMPLES * 2) & !7);
             memory_barrier();
 
             // Start DMA
